// src/types/scenes.ts
export interface PlayerCharacterConfig {
  /** Parent Phaser scene */
  scene: Phaser.Scene;
  
  /** Initial X position */
  x: number;
  
  /** Initial Y position */
  y: number;
  
  /** Sprite asset key to use for player visual */
  spriteKey: string;
  
  /** Movement speed in pixels per second (default: 150) */
  speed?: number;
}
// ...
export interface WASDKeys {
  up: Phaser.Input.Keyboard.Key;
  down: Phaser.Input.Keyboard.Key;
  left: Phaser.Input.Keyboard.Key;
  right: Phaser.Input.Keyboard.Key;
}
// ...
export interface Position {
  x: number;
  y: number;
}